Basic Autocrypt Integration
Process headers and look up keys in Autocrypt Store.
This is a combined commit of what was discussed in #654.

Process headers
------------------

Api: keyring.processHeader will store keys from Autocrypt headers.
Compare addr part of header to plain email

processHeader will take a headers object with the Autocrypt, from and
date headers.

Reject autocrypt headers that are > 10k.
Corresponding PR to the Autocrypt spec is at
  https://github.com/autocrypt/autocrypt/pull/397

Keep state
-------------

Use mvelo storage to persist state.
Use distinct autocrypt stores for different identities
Store all autocrypt data for a given keyring in one mvelo.storage entry.
prefix is mailvelope.autocrypt

Introduce a (private) Store class on top of mvelo.storage.
A store uses a Map internally but serializes it as an object.
It has a callback based API so we can use it with Autocrypt.

Look up keys
---------------

Rename autoLocate to keyRegistry and pass keyring id as param.
Storage and Lookup of keys is based on current identity
so that autocrypt keys do not leak from one identity to another.

Search for public keys uses:
* Mailvelope Key Server (MKS)
* web key directory (WKD)
* Autocrypt (AC)

It will return the best key it finds and report the source.
keyRegistry.lookup returns an Object not just a key
The object has content (the armored key) and source properties.

The sources also return a timestamp to compare to results from other sources.
For now we will keep using the first strategy that returns anything.
But the logic may become more evolved in the future.
Prepare the (internal) lookup api for that.

Use Autocrypt js
----------------

Introduce autocryptWrapper (For the lack of a better name...)
It wraps the storage and the callback based calls
to autocrypt-js Autocrypt class in an asynchronous API.

Add autocrypt flag in settings similar to wkd and mailvelope key server.

API integration of Autocrypt
----------------------------

* handle the process-autocrypt-header event in the clientAPI
  It will trigger the event in the apiController
* convert the date string to an actual date
* there is a whitelist for attribute types for events.
  added date, fromAddr and header to expect strings.

Also catching Error('Invalid Autocrypt Header: no valid header found').
If there is no autocrypt header at all that should not cause an error.

Other
-----

doc: upcase {string} when used as a type in api docs
locales: add german translation for autocrypt
mention Autocrypt in the message of the en source.