Initial test for the API
API level tests allow testing the wiring in the event handling
inside the controllers and the content scripts.
It ensures that API calls actually accomplish their goals.

test/polyfills.js now includes a simple implementation of
sendMessage that triggers the corresponding functions in the main
controller (which in turn calls handleApiEvent in the api controller).

The local storage is stubbed entirely so we do not have to worry
about replacing it with rewire in some contexts.
We do not recover it after the run because there is no
working local storage in the test environment anyway.