make keyring.init clear internal state
We keep the keyringMap and keyringAttr as constants in the keyring closure.
This way they act as singletons that get initialized at load time
and then they keep some state that is also synchronized to the store.

During normal runs this is fine. But in tests the state in keyringMap
and keyringAttr gets out of sync if we clear the locale storage.
This way test state leaks from one test to the next.

The keyring test even made use of this creating a keyring in one test
and then depending on it in another.

This change clears the internal state of the keyring on init.

The dependency between the tests has also been removed
and we now have both tests run based on the same data
created during the initial beforeEach hooks.